LayoutTests/fast/js/cached-call-uninitialized-arguments-expected.txt

This test checks that uninitialized parameters for cached call functions correctly defaults to undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS arg4 is undefined.
PASS arg5 is undefined.
PASS arg6 is undefined.
PASS successfullyParsed is true

TEST COMPLETE
